requirements.txt
Django==3.2.9
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
mysqlclient==2.1.0
